Terminado módulo de Estructuración de API y documentación con Swagger

+++ src/main/java/co/edu/platzi/spring/persistence/entity/Cliente.java
@@ -25,7 +25,7 @@ public class Cliente {
 	
 	private String apellidos;
 	
-	private Integer celular;
+	private Long celular;
 	
 	private String direccion;
